photos update in database and shared library
