+++ MagazinE.sln
@@ -9,9 +9,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MagazinEAPI", "MagazinEAPI\
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "UnitTests", "UnitTests\UnitTests.csproj", "{531C447D-43BA-48DB-B93C-2CAFCCBC56BC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebsiteFrontMagazinE.Server", "WebsiteFrontMagazinE\WebsiteFrontMagazinE.Server\WebsiteFrontMagazinE.Server.csproj", "{969E1966-2722-4730-8AA9-39638AF0F0E7}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebsiteFrontMagazinE.Server", "WebsiteFrontMagazinE\WebsiteFrontMagazinE.Server\WebsiteFrontMagazinE.Server.csproj", "{969E1966-2722-4730-8AA9-39638AF0F0E7}"
 EndProject
 Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "websitefrontmagazine.client", "WebsiteFrontMagazinE\websitefrontmagazine.client\websitefrontmagazine.client.esproj", "{4CB1B653-8C62-4888-A113-4CBE8F40E575}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedLibrary", "SharedLibrary\SharedLibrary.csproj", "{E333E1E1-ED84-4ACF-9491-40B409043190}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -41,6 +43,10 @@ Global
 		{4CB1B653-8C62-4888-A113-4CBE8F40E575}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4CB1B653-8C62-4888-A113-4CBE8F40E575}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4CB1B653-8C62-4888-A113-4CBE8F40E575}.Release|Any CPU.Deploy.0 = Release|Any CPU
+		{E333E1E1-ED84-4ACF-9491-40B409043190}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E333E1E1-ED84-4ACF-9491-40B409043190}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E333E1E1-ED84-4ACF-9491-40B409043190}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E333E1E1-ED84-4ACF-9491-40B409043190}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
