+++ MagazinEAPI/Models/Ban.cs
@@ -1,16 +1,6 @@
-﻿namespace MagazinEAPI.Models
+﻿using SharedLibrary.Base_Classes___Database;
+namespace MagazinEAPI.Models
 {
-	public abstract class BanAbstract
-	{
-		public int Id { get; set; }
-		public int AdminId { get; set; }
-		public int UserId { get; set; }
-	}
-
-	public class BanDTO: BanAbstract
-	{
-	}
-
 	public class Ban : BanAbstract
 	{
 		public Admin Admin { get; set; } = null!;
