+++ MagazinEAPI/Models/HeadEditor.cs
@@ -1,13 +1,5 @@
 ﻿namespace MagazinEAPI.Models
 {
-
-	public class HeadEditorDTO : RoleDTO
-	{
-		public List<int> ArticlesIds { get; set; }
-		public List<int> EditorsUnderIds { get; set; }
-		public List<int> JournalistsUnderIds { get; set; }
-	}
-
 	public class HeadEditor : Role
 	{
 		//HeadEditor get articles from their journalists
