+++ MagazinEAPI/Models/PhotoArticle.cs
@@ -1,0 +1,10 @@
+﻿namespace MagazinEAPI.Models
+{
+	public class PhotoArticle
+	{
+		public int PhotoId { get; set; }
+		public int ArticleId { get; set; }
+		public Photo Photo { get; set; } = null!;
+		public Article Article { get; set; } = null!;
+	}
+}
