+++ MagazinEAPI/Models/PublishRequest.cs
@@ -1,22 +1,9 @@
-﻿namespace MagazinEAPI.Models
+﻿using SharedLibrary.Base_Classes___Database;
+namespace MagazinEAPI.Models
 {
-	public abstract class PublishRequestAbstract
-	{
-		public int Id { get; set; }
-		public DateTime Date { get; set; }
-		public int ArticleId { get; set; }
-		public PublishState PublishState { get; set; }
-	}
-
-	public class PublishRequestDTO : PublishRequestAbstract
-	{
-		public int ReviewerId { get; set; }
-
-	}
-
 	public class PublishRequest : PublishRequestAbstract
 	{
-		public Article Article { get; set; } = null!; // has 1 Article
+		public Article Article { get; set; } = null!; // has 1 Article -> ArticleId in abstractClass
 
 
 		public int? ReviewerId { get; set; }
