--- MagazinEAPI/Models/RegisterRequestcs.cs
@@ -1,11 +1,0 @@
-﻿namespace MagazinEAPI.Models
-{
-	public class RegisterRequest
-	{
-		public int Id { get; set; }
-		public string FirstName { get; set; }
-		public string LastName { get; set; }
-		public string Email { get; set; }
-		public string Password { get; set; }
-	}
-}
