+++ MagazinEAPI/Models/Role.cs
@@ -1,16 +1,6 @@
-﻿namespace MagazinEAPI.Models
+﻿using SharedLibrary.Base_Classes___Database;
+namespace MagazinEAPI.Models
 {
-	public abstract class RoleBase
-	{
-		public int Id { get; set; }
-	}
-
-
-	public abstract class RoleDTO : RoleBase
-	{
-		public int PersonInfoId { get; set; }
-	}
-
 	public abstract class Role : RoleBase
 	{
 		public string ApplicationUserId { get; set; } = null!;
