+++ MagazinEAPI/Models/UnbanRequest.cs
@@ -1,16 +1,6 @@
-﻿namespace MagazinEAPI.Models
+﻿using SharedLibrary.Base_Classes___Database;
+namespace MagazinEAPI.Models
 {
-	public abstract class UnbanRequestAbstract
-	{
-		public int Id { get; set; }
-		public string Reason { get; set; }
-		public int BanId { get; set; }
-		public UnbanRequestState State { get; set; }
-	}
-	public class UnbanRequestDTO : UnbanRequestAbstract
-	{
-		
-	}
 
 	public class UnbanRequest : UnbanRequestAbstract
 	{
