+++ SharedLibrary/Base Classes - Database/RegisterRequest.cs
@@ -1,0 +1,10 @@
+﻿namespace SharedLibrary.Base_Classes___Database;
+
+public class RegisterRequest
+{
+	public int Id { get; set; }
+	public string FirstName { get; set; }
+	public string LastName { get; set; }
+	public string Email { get; set; }
+	public string Password { get; set; }
+}
