+++ SharedLibrary/Base Classes - Database/RoleBase.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SharedLibrary.Base_Classes___Database
+{
+	public abstract class RoleBase
+	{
+		public int Id { get; set; }
+	}
+}
