+++ SharedLibrary/Base Classes - Database/TagAbstract.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SharedLibrary.Base_Classes___Database
+{
+	public abstract class TagAbstract
+	{
+		public int Id { get; set; }
+		public string Name { get; set; }
+	}
+}
