+++ SharedLibrary/DTO Classes/PublishRequestDTO.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using SharedLibrary.Base_Classes___Database;
+
+namespace SharedLibrary.DTO_Classes
+{
+	public class PublishRequestDTO : PublishRequestAbstract
+	{
+		public int ReviewerId { get; set; }
+
+	}
+}
