+++ SharedLibrary/DTO Classes/RoleDTO.cs
@@ -1,0 +1,8 @@
+﻿using SharedLibrary.Base_Classes___Database;
+namespace SharedLibrary.DTO_Classes
+{
+	public abstract class RoleDTO : RoleBase
+	{
+		public string PersonInfoId { get; set; }
+	}
+}
