+++ SharedLibrary/DTO Classes/UnbanRequestDTO.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using SharedLibrary.Base_Classes___Database;
+
+namespace SharedLibrary.DTO_Classes
+{
+	public class UnbanRequestDTO : UnbanRequestAbstract
+	{
+
+	}
+}
